+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `VerticalStepper`.
-- Feature: Add `WarningIcon`.
+- Feature: Add `WarningIcon` and `WaitingIcon`.
 - Feature: Update `CheckedIcon` to accept `color` and `title` props.
 - Feature: Add a new `Button` component.